pos_backend/requirements.txt
Django==4.2.10
djangorestframework==3.15.1
mysqlclient==2.2.4
django-cors-headers==4.3.1
Pillow==10.2.0
python-decouple==3.8
